Add integer size scaling to coerceDType
- Actually treat the expected type as a minimum as is true in nwb-schema.
If the value type is larger than the expected type, then scale progressively to fit.

- Fix dynamic table indexing due to VectorIndex now being set to uint8 as an expected
type. The type should start at uint8 but scale upwards as the indices get larger.

+++ +types/+util/+dynamictable/addRawData.m
@@ -34,16 +34,16 @@ if ~isempty(index)
     
     raggedOffset = 0;
     if isa(VecInd.data, 'types.untyped.DataPipe') && 0 < VecInd.data.dims
-        raggedOffset = VecInd.data.load(VecInd.data.dims);
+        raggedOffset = double(VecInd.data.load(VecInd.data.dims));
     elseif ~isempty(VecInd.data)
-        raggedOffset = VecInd.data(end);
+        raggedOffset = double(VecInd.data(end));
     end
     
     raggedValue = raggedOffset + size(data, 1);
     if isa(VecInd.data, 'types.untyped.DataPipe')
         VecInd.data.append(raggedValue);
     else
-        VecInd.data = [VecInd.data; raggedValue];
+        VecInd.data = [double(VecInd.data); raggedValue];
     end
 end
